BoneNormalizer.cs
