+++ Assets/VRM/UniVRM/Editor/Format/VRMHumanoidNormalizerMenu.cs
@@ -49,7 +49,7 @@ namespace VRM
             var go = Selection.activeObject as GameObject;
 
             // BoneNormalizer.Execute はコピーを正規化する。UNDO無用
-            Selection.activeGameObject = BoneNormalizer.Execute(go, true, false);
+            Selection.activeGameObject = VRMBoneNormalizer.Execute(go, true, false);
         }
     }
 }
